# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for the research monitor (arXiv ingestion + paper catalogue).",
        "title": "Research Monitor API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/arxiv/downloads/{job_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a snapshot of a PDF-download job scheduled by\n/api/arxiv/fetch, including per-entry results and totals.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDFDownload"
                ],
                "summary": "Get the status of a PDF-download job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Download job id (UUIDv4)",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job status snapshot",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_paper.JobStatusEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Job unknown or expired",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/arxiv/downloads/{job_id}/stream": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Opens a Server-Sent Events stream for a PDF-download\njob: replays buffered events first, then forwards live\nprogress events and a terminal summary frame.",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "PDFDownload"
                ],
                "summary": "Stream PDF-download events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Download job id (UUIDv4)",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE stream of download.progress events followed by a terminal download.summary",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Job unknown or expired",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/papers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Papers"
                ],
                "summary": "List all persisted papers",
                "responses": {
                    "200": {
                        "description": "Paper catalogue",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_paper.PaperListEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Catalogue unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/papers/{source}/{source_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Papers"
                ],
                "summary": "Get a paper by source + source_id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source identifier (e.g. 'arxiv')",
                        "name": "source",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Source-local paper id (e.g. arXiv id)",
                        "name": "source_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paper found",
                        "schema": {
                            "$ref": "#/definitions/internal_http_controller_paper.PaperEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API token",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Paper not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Catalogue unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_yoavweber_research-monitor_backend_internal_http_common.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_yoavweber_research-monitor_backend_internal_http_common.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_yoavweber_research-monitor_backend_internal_http_common.Error"
                }
            }
        },
...
        "internal_http_controller_paper.DownloadEntryResultDTO": {
            "type": "object",
            "properties": {
                "bytes": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "completed_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "paper_id": {
                    "$ref": "#/definitions/internal_http_controller_paper.PaperIDDTO"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_http_controller_paper.DownloadJobSnapshotDTO": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "completed_at": {
                    "type": "string"
                },
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_http_controller_paper.DownloadEntryResultDTO"
                    }
                },
                "failed": {
                    "type": "integer"
                },
                "job_id": {
                    "type": "string"
                },
                "succeeded": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_http_controller_paper.JobStatusEnvelope": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_http_controller_paper.DownloadJobSnapshotDTO"
                }
            }
        },
        "internal_http_controller_paper.PaperEnvelope": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_http_controller_paper.PaperResponse"
                }
            }
        },
        "internal_http_controller_paper.PaperIDDTO": {
            "type": "object",
            "properties": {
                "source": {
                    "type": "string"
                },
                "source_id": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "internal_http_controller_paper.PaperListEnvelope": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_http_controller_paper.PaperListResponse"
                }
            }
        },
        "internal_http_controller_paper.PaperListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "papers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_http_controller_paper.PaperResponse"
                    }
                }
            }
        },
        "internal_http_controller_paper.PaperResponse": {
            "type": "object",
            "properties": {
                "abs_url": {
                    "type": "string"
                },
                "abstract": {
                    "type": "string"
                },
                "authors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "pdf_url": {
                    "type": "string"
                },
                "primary_category": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "source_id": {
                    "type": "string"
                },
                "submitted_at": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT issued by POST /auth/login. Supply as \"Bearer \u0026lt;token\u0026gt;\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}